+++ benchmarks/BUILD.bazel
@@ -192,8 +192,8 @@ cc_binary(
         "//src/option/table:spliced_surface_builder",
         "//src/option/table:eep_transform",
         "//src/math:bspline_nd_separable",
-        "//src/option/table:dimensionless_builder",
-        "//src/option/table:dimensionless_inner",
+        "//src/option/table/dimensionless:dimensionless_builder",
+        "//src/option/table/dimensionless:dimensionless_inner",
         "@google_benchmark//:benchmark",
     ],
     tags = ["benchmark", "manual"],
@@ -477,8 +477,8 @@ cc_binary(
         "//src/option:american_option_batch",
         "//src/option:iv_solver",
         "//src/option:interpolated_iv_solver",
-        "//src/option/table:dimensionless_builder",
-        "//src/option/table:dimensionless_inner",
+        "//src/option/table/dimensionless:dimensionless_builder",
+        "//src/option/table/dimensionless:dimensionless_inner",
     ],
     tags = ["manual"],
 )
