+++ benchmarks/interp_iv_safety.cc
@@ -28,8 +28,8 @@
 #include <string>
 #include <vector>
 
-#include "mango/option/table/dimensionless_builder.hpp"
-#include "mango/option/table/dimensionless_inner.hpp"
+#include "mango/option/table/dimensionless/dimensionless_builder.hpp"
+#include "mango/option/table/dimensionless/dimensionless_inner.hpp"
 
 using namespace mango;
 using namespace mango::bench;
