+++ src/option/table/dimensionless/dimensionless_adaptive.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
-#include "mango/option/table/dimensionless_builder.hpp"
-#include "mango/option/table/dimensionless_european.hpp"
+#include "mango/option/table/dimensionless/dimensionless_builder.hpp"
+#include "mango/option/table/dimensionless/dimensionless_european.hpp"
 #include "mango/option/table/error_attribution.hpp"
 #include "mango/option/american_option_batch.hpp"
 #include "mango/math/cubic_spline_solver.hpp"
