+++ src/option/table/dimensionless/golden_surface_data.inc
