+++ tests/dimensionless_adaptive_test.cc
@@ -4,8 +4,8 @@
  * @brief Tests for adaptive dimensionless 3D surface builder
  */
 
-#include "mango/option/table/dimensionless_builder.hpp"
-#include "mango/option/table/dimensionless_european.hpp"
+#include "mango/option/table/dimensionless/dimensionless_builder.hpp"
+#include "mango/option/table/dimensionless/dimensionless_european.hpp"
 #include "mango/option/american_option_batch.hpp"
 #include "mango/math/cubic_spline_solver.hpp"
